Clean up complex field hierarchy to improve readability
