+++ lib/Alchemy/Phrasea/SearchEngine/Elastic/Mapping/ComplexFieldMapping.php
@@ -13,72 +13,12 @@ namespace Alchemy\Phrasea\SearchEngine\Elastic\Mapping;
 
 use Alchemy\Phrasea\SearchEngine\Elastic\FieldMapping;
 
-class ComplexFieldMapping extends FieldMapping
+class ComplexFieldMapping extends ComplexMapping
 {
-    /**
-     * @var FieldMapping[]
-     */
-    private $children = [];
 
-    /**
-     * @var null|string
-     */
-    private $childKey = 'fields';
-
-    public function useAsPropertyContainer()
+    public function __construct($name, $type = null)
     {
-        $this->childKey = 'properties';
-    }
-
-    public function useAsFieldContainer()
-    {
-        $this->childKey = 'fields';
-    }
-
-    public function useAsBareContainer()
-    {
-        $this->childKey = null;
-    }
-
-    /**
-     * @param FieldMapping $child
-     * @return FieldMapping
-     */
-    public function addChild(FieldMapping $child)
-    {
-        if (isset($this->children[$child->getName()])) {
-            throw new \LogicException(sprintf('There is already a "%s" multi field.', $child->getName()));
-        }
-
-        if ($child->getType() !== $this->getType() && $this->getType() !== self::TYPE_OBJECT) {
-            throw new \LogicException('Child field type must match parent type.');
-        }
-
-        return $this->children[$child->getName()] = $child;
-    }
-
-    /**
-     * @return RawFieldMapping
-     */
-    public function addRawChild()
-    {
-        return $this->addChild(new RawFieldMapping($this->getType()));
-    }
-
-    /**
-     * @return bool
-     */
-    public function hasChildren()
-    {
-        return ! empty($this->children);
-    }
-
-    /**
-     * @return FieldMapping[]
-     */
-    public function getChildren()
-    {
-        return $this->children;
+        parent::__construct($name, $type ?: FieldMapping::TYPE_OBJECT);
     }
 
     /**
@@ -86,20 +26,6 @@ class ComplexFieldMapping extends FieldMapping
      */
     protected function getProperties()
     {
-        if (! $this->hasChildren()) {
-            return [];
-        }
-
-        $properties = [ ];
-
-        foreach ($this->children as $name => $child) {
-            $properties[$name] = $child->toArray();
-        }
-
-        if ($this->childKey) {
-            return [$this->childKey => $properties];
-        }
-
-        return $properties;
+        return [ 'fields' => parent::getProperties() ];
     }
 }
